# Remove abc inheritance from Serializable