@vyasr 2021-05-14
Here are some back of the envelope benchmarks run using IPython's `timeit` magic. Note that the last three rows (creating a `cudf.Index`) are in microseconds, whereas the first six (simple `isinstance` checks) are in nanoseconds.

|                                | Old               | New              |
|--------------------------------|-------------------|------------------|
| isinstance(1, cudf.DataFrame)  | 468 ns ± 8.86 ns  | 133 ns ± 1.53 ns |
| isinstance(1, cudf.Series)     | 459 ns ± 3.12 ns  | 134 ns ± 3.08 ns |
| isinstance(1, cudf.Index       | 480 ns ± 0.889 ns | 129 ns ± 2.3 ns  |
| isinstance(df, cudf.DataFrame) | 115 ns ± 0.236 ns | 129 ns ± 1.06 ns |
| isinstance(ser, cudf.Series)   | 772 ns ± 3.3 ns   | 156 ns ± 1.33 ns |
| isinstance(idx, cudf.Index)    | 776 ns ± 16.1 ns  | 155 ns ± 2.88 ns |
| cudf.Index(df)                 | 347 µs ± 1.64 µs  | 307 µs ± 5.09 µs |
| cudf.Index(ser)                | 416 µs ± 205 ns   | 388 µs ± 3.59 µs |
| cudf.Index(idx)                | 417 µs ± 1.2 µs   | 386 µs ± 1.53 µs |