@jakirkham 2021-05-18
Just to note, cuML uses `Serializable` via [`Buffer`]( https://github.com/rapidsai/cuml/blob/4d069915b9d12fa8b301d1c5e4bb4abe5198d1a4/python/cuml/common/array.py#L33 )

cc @dantegd