@jakirkham 2021-05-18
Are there use cases where this `isinstance` overhead is noticeable? Do we have an end user workflow where making this change has an impact?