@vyasr 2021-05-18
> Are there use cases where this `isinstance` overhead is noticeable? Do we have an end user workflow where making this change has an impact?

Yes, unfortunately. In an ideal world we would rely on ducktyping everywhere, but that's definitely not the case internally for various reasons. As a result, lots of different code paths are impacted here. I included a benchmark above of the difference in constructing an Index from different types, here's the relevant data again:

|                                | Old               | New              |
|--------------------------------|-------------------|------------------|
| cudf.Index(df)                 | 347 µs ± 1.64 µs  | 307 µs ± 5.09 µs |
| cudf.Index(ser)                | 416 µs ± 205 ns   | 388 µs ± 3.59 µs |
| cudf.Index(idx)                | 417 µs ± 1.2 µs   | 386 µs ± 1.53 µs |

That was just intended as a representative example. The same type of difference will be observed in a large number of different operations. Take this simple binop example (run via IPython):
```
In [1]: import cudf
In [2]: x = cudf.Series([1, 2, 3])
In [3]: %timeit x + x
```

The before/after numbers there are `158 µs ± 961 ns`/`139 µs ± 619 ns per loop`. If I change that to a `Series` of size 1 million instead of 3, the difference is comparable (as expected): `531 µs ± 5.86 µs`/`508 µs ± 4.17 µs`. Obviously these aren't massive changes, but they can add up when the effect is so pervasive throughout the code base. As long as we're refactoring the code with performance in mind we may as well avoid introducing unnecessary overheads like this.

In addition, removing `ABCMeta` from the metaclass hierarchy of all our classes makes it simpler to add custom metaclasses if necessary. Custom metaclasses are another thing that I'd love to avoid, but there's at least one case where matching `pandas` semantics requires us to do some things that would be more cleanly accomplished using a custom metaclass than by any other approach (including what we currently do) and it's easier to not have to worry about how overriding methods like `__instancecheck__` will behave when creating a metaclass that inherits from `ABCMeta`.